feat(management): add the possibility to store others kinds of IdP
Closes gravitee-io/issues#3155

+++ src/management/configuration/identity/identity-providers.component.ts
@@ -19,7 +19,8 @@ import IdentityProviderService from '../../../services/identityProvider.service'
 import NotificationService from '../../../services/notification.service';
 import PortalConfigService from '../../../services/portalConfig.service';
 import _ = require('lodash');
-import {IScope} from 'angular';
+import {IScope, material} from 'angular';
+import UserService from '../../../services/user.service';
 
 const IdentityProvidersComponent: ng.IComponentOptions = {
   bindings: {
@@ -31,6 +32,7 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
     IdentityProviderService: IdentityProviderService,
     PortalConfigService: PortalConfigService,
     NotificationService: NotificationService,
+    UserService: UserService,
     $state: StateService,
     Constants,
     $rootScope: IScope
@@ -39,37 +41,53 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
     this.$rootScope = $rootScope;
     this.settings = _.cloneDeep(Constants);
 
+    this.currentUserSource = UserService.currentUser.source;
     this.availableProviders = [
       {'name': 'Gravitee.io AM', 'icon': 'perm_identity', 'type': 'graviteeio_am'},
       {'name': 'Google', 'icon': 'google-plus', 'type': 'google'},
       {'name': 'GitHub', 'icon': 'github-circle', 'type': 'github'},
+      {'name': 'LDAP', 'icon': 'group', 'type': 'ldap'},
       {'name': 'OpenID Connect', 'icon': 'perm_identity', 'type': 'oidc'}
     ];
+
+    this.$onInit = () => {
+      this.socialIdentityProviders = _.filter(this.identityProviders, idp => idp.type === 'github' || idp.type === 'google' || idp.type === 'graviteeio_am' || idp.type === 'oidc');
+      this.internalIdentityProviders = _.filter(this.identityProviders, idp => idp.type === 'memory' || idp.type === 'gravitee' || idp.type === 'ldap');
+    };
 
     this.create = (type) => {
       $state.go('management.settings.identityproviders.new', {type: type});
     };
 
     this.delete = (provider: IdentityProvider) => {
-      let that = this;
-      $mdDialog.show({
-        controller: 'DialogConfirmController',
-        controllerAs: 'ctrl',
-        template: require('../../../components/dialog/confirmWarning.dialog.html'),
-        clickOutsideToClose: true,
-        locals: {
-          title: 'Are you sure you want to delete this identity provider ?',
-          msg: '',
-          confirmButton: 'Delete'
-        }
-      }).then(function (response) {
-        if (response) {
-          IdentityProviderService.delete(provider).then(response => {
-            NotificationService.show('Identity provider \'' + provider.name + '\' has been deleted');
-            $state.go('management.settings.identityproviders.list', {}, {reload: true});
-          });
-        }
-      });
+      if (this.currentUserSource === provider.id) {
+        var alert = $mdDialog.alert()
+          .title('Attention')
+          .textContent('Deleting this identity provider is forbidden since you have been authenticated with')
+          .ok('Close');
+
+        $mdDialog.show(alert);
+      } else {
+        let that = this;
+        $mdDialog.show({
+          controller: 'DialogConfirmController',
+          controllerAs: 'ctrl',
+          template: require('../../../components/dialog/confirmWarning.dialog.html'),
+          clickOutsideToClose: true,
+          locals: {
+            title: 'Are you sure you want to delete this identity provider ?',
+            msg: '',
+            confirmButton: 'Delete'
+          }
+        }).then(function (response) {
+          if (response) {
+            IdentityProviderService.delete(provider).then(response => {
+              NotificationService.show('Identity provider \'' + provider.name + '\' has been deleted');
+              $state.go('management.settings.identityproviders.list', {}, {reload: true});
+            });
+          }
+        });
+      }
     };
 
     this.saveForceLogin = () => {
@@ -96,6 +114,52 @@ const IdentityProvidersComponent: ng.IComponentOptions = {
         NotificationService.show('Login form is now ' + (this.settings.authentication.localLogin.enabled ? 'enabled' : 'disabled'));
         Constants.authentication.localLogin =  response.data.authentication.localLogin;
       });
+    };
+
+    this.changeIdentityProviderStatus = (identityProvider: IdentityProvider, newStatus: boolean) => {
+      if (this.currentUserSource === identityProvider.id) {
+        var alert = $mdDialog.alert()
+          .title('Attention')
+          .textContent('Disabling this identity provider is forbidden since you have been authenticated with')
+          .ok('Close');
+
+        $mdDialog.show(alert);
+      } else {
+        IdentityProviderService.get(identityProvider.id).then(response => {
+          let idp: IdentityProvider = response;
+          idp.enabled = newStatus;
+          IdentityProviderService.update(idp).then(response => {
+            NotificationService.show('Identity provider \'' + idp.name + '\' has been updated');
+            identityProvider.enabled = newStatus;
+          });
+        });
+      }
+    };
+
+    this.upward = (identityProvider: IdentityProvider) => {
+      IdentityProviderService.get(identityProvider.id).then(response => {
+        let idp: IdentityProvider = response;
+        idp.order = idp.order - 1;
+        IdentityProviderService.update(idp).then(response => {
+          NotificationService.show('Identity provider \'' + idp.name + '\' order has been changed with success');
+          this.refresh();
+        });
+      });
+    };
+
+    this.downward = (identityProvider: IdentityProvider) => {
+      IdentityProviderService.get(identityProvider.id).then(response => {
+        let idp: IdentityProvider = response;
+        idp.order = idp.order + 1;
+        IdentityProviderService.update(idp).then(response => {
+          NotificationService.show('Identity provider \'' + idp.name + '\' order has been changed with success');
+          this.refresh();
+        });
+      });
+    };
+
+    this.refresh = () => {
+      IdentityProviderService.list().then(response => this.identityProviders = response);
     };
   }
 };
